Commit of fixture - dummy users who simulate yearbook ranking configuration. Yearbook ranking code complete. Next: test ranking class.

+++ voomza/apps/yearbook/ajax.py
@@ -1,12 +1,13 @@
 import json, logging
 from dajaxice.decorators import dajaxice_register
 from django.core.urlresolvers import reverse
-from django.http import HttpResponseForbidden
+from django.http import HttpResponseForbidden, HttpResponseNotFound
 from django_facebook.api import require_persistent_graph
 from voomza.apps.account.models import YearbookFacebookUser
 from voomza.apps.yearbook.api import YearbookFacebookUserConverter
 from yearbook.tasks import get_and_store_top_friends_fast, get_optional_profile_fields
-from voomza.apps.yearbook.models import InviteRequestSent, Badge, BadgeVote
+from voomza.apps.yearbook.models import InviteRequestSent, Badge, BadgeVote, YearbookSign
+from voomza.apps.yearbook.ranking import UserProfileRanking
 
 
 logger = logging.getLogger(name=__name__)
@@ -99,6 +100,9 @@ def save_badge_votes(request, selected_friends, next_view='sign_friends'):
     Saves the friends this user has indicated
     are family, friends, etc.
     """
+    if not request.user.is_authenticated():
+        return HttpResponseForbidden
+
     # If for some reason they have some in the db, clear them
     old_votes = BadgeVote.objects.filter(from_user=request.user)
     old_votes.delete()
@@ -112,10 +116,86 @@ def save_badge_votes(request, selected_friends, next_view='sign_friends'):
     return send_to_next_page(request.user, next_view)
 
 
+YEARBOOKS_PER_PAGE = 6
+
 @dajaxice_register
-def get_yearbooks_to_sign(request):
+def get_yearbooks_to_sign(request, offset=0):
     """
     Returns a list of users that we suggest this
     user should sign.
     """
-    pass
+    if not request.user.is_authenticated():
+        return HttpResponseForbidden
+
+    if 'all_yearbooks' in request.session:
+        all_yearbooks = request.session['all_yearbooks']
+    else:
+        profile_ranking = UserProfileRanking(request.user.profile)
+        all_yearbooks = profile_ranking.get_yearbooks_to_sign()
+        request.session['all_yearbooks'] = all_yearbooks
+
+    # Stuff in the list is either User or YearbookFacebookUser
+    # we need fields name, pic and facebook_id
+    return json.dumps([{
+        #   | YearbookFacebookUser ||         User         |
+        'name': getattr(x, 'name') or x.profile.facebook_name,
+        'pic':  getattr(x, 'pic_square') or x.profile.pic_square,
+        'id': getattr(x, 'facebook_id') or x.profile.facebook_id,
+    } for x in all_yearbooks[offset:offset+YEARBOOKS_PER_PAGE]])
+
+
+@dajaxice_register
+def save_yearbook_sign(request, text, to_id):
+    """
+    Saves a user's message, note that
+    the person being signed may or may not be
+    signed up for the app
+    """
+    if not request.user.is_authenticated():
+        return HttpResponseForbidden
+    sign = YearbookSign(from_user = request.user, to_id=to_id, text=text)
+    sign.save()
+    return json.dumps('ok')
+
+
+SIGNS_PER_PAGE = 10
+
+@dajaxice_register
+def get_yearbook_signs(request, offset=0):
+    """
+    Returns the list of users who have already
+    signed this person's yearbook
+    """
+    if not request.user.is_authenticated():
+        return HttpResponseForbidden
+
+    signs_query = YearbookSign.objects.filter(to_id=request.user.profile.facebook_id)
+    signs = signs_query[offset:offset+SIGNS_PER_PAGE].values(
+        'id', 'from_user__profile__name', 'from_user__profile__pic_square'
+    )
+
+    import ipdb
+    ipdb.set_trace()
+
+    # Serialize and return
+    # Return the offset so the caller can reject duplicates
+    return json.dumps({
+        'signs': list(signs),
+        'offset': offset,
+    })
+
+
+@dajaxice_register
+def get_yearbook_sign_message(request, sign_id):
+    """
+    Returns the list of users who have already
+    signed this yearbook
+    """
+    if not request.user.is_authenticated():
+        return HttpResponseForbidden
+
+    try:
+        sign = YearbookSign.objects.get(id=sign_id).values('text')
+        return json.dumps(sign)
+    except YearbookSign.DoesNotExist:
+        return HttpResponseNotFound
